+++ HISTORY.md
@@ -3,10 +3,14 @@
 ## in development for v7:
 * **New Features:**
     * Scenes
+    * unified blocks palette option, thanks, Michael!
 * **Notable Changes:**
     * saved projects remember the last edited srpite
 * **Notable Fixes:**
     * made scrollbars in the wardrobe and jukebox more responsive
+
+### 2021-07-02
+* gui, object, store, etc.: unified blocks palette option, thanks, Michael!
 
 ### 2021-05-21
 * gui, scenes, store: proxied thumbnail, name and notes in project, restored in XML